pyproject.toml
[project]
name = "slackbolt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "python-decouple>=3.8",
    "requests>=2.32.4",
    "slack-bolt>=1.23.0",
]

[tool.pytest.ini_options]
addopts = "--ignore=tests/integration"
